Add files via upload
Stage 1 Objectives:
Import tensorflow and numpy to your program. The first one loads the data, the second one transforms it;
Load the data in your program. You need x_train and y_train only. Skip x_test and y_test in return of load_data(), we will create them ourselves in the next stage.
Sometimes x_train or x_test will be called the features array (because they contain brightnesses of the pixels, which are the images' features). y_train or y_test will be called the target array (because they contain classes, digits which we are going to predict);
Reshape the features array to the 2D array with 
n
n
 rows (
n
n
 = number of images in the dataset) and 
m
m
 columns (
m
m
 = number of pixels in each image);
Print information about the dataset: target classes' names; the shape of the features array; the shape of the target array; the minimum and maximum values of the features array. Use the following format for the shapes: (the number of rows, the number of columns)

EXAMPLE OUTPUT:

Classes: [1 2 3 5 6]
Features' shape: (200, 100)
Target's shape: (200,)
min: 0.0, max: 100.0